Update .gitignore file using Github templates.
The original .gitignore was copied from the initial commit in the
binarystore repo. This was based on Github public template for autotools
with additions by the author for that project. The original template
(linked in the .gitignore) has been updated, so now it reflects those
changes. I also concated the C++ gitignore template, which covers most
of the additions originally made.

For future reference, I ran `wget <raw template urls> -O .gitignore` and
then manually added the source URLs to the file, in addition to rules
specific to this repo.

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: I390c0a8f79466580b4b9cc11d7a13cbd0b6309e9