Add tests for RrdService::handle method delegation.
Currently, tests only check that each method (i.e. last) is capable of
properly handling a request and that handle is able to signal an error
if there is an unparsable buffer or a request with an unset method. This
change uses a RrdService mock to check the interaction between the
handle method and the service object.

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: I121102cf9ff87b12b29051627de9a104ee9b4169