Add RRD service exposing a protobuf interface.
Essentially the glue between the IPMI handler and librrdplus, this
component will take messages with a request and translate them into a
library call, packing the result into another proto message.

I based the API on the protobuf generic_services[1], since the handler
essentially implements a poor man's RPC and someone seemed to think that
this is a good, flexible approach. However, I left out the callback
since async operation is not within scope (yet).

[1] https://developers.google.com/protocol-buffers/docs/reference/cpp-generated#service

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: I2636c23ac77c36fe874a75e3f2d964a56de9bca3