Fix RrdBlobHandler::open path check and log messages.
Although the IPMI Blob handler library specifies that commands should
only be called with paths that the handler canHandle, this check was not
explicit on the standalone class RrdBlobHandler open method. This method
also incorrectly logged a successful open when the session id already
existed. This change fixes both of this issues.

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: I0d713afcfa905a606e170a32cb75e8cb7a052d0a