Implement expire command to close a session.
Since the handler implements a maximum number of open sessions, if the
client forgets or is unable to close a session (program error, crash, or
IPMI drops a message), then the handler might end up blocked. In order
to provide a more graceful recovery, this change implements expire which
seems to be used by the IPMI Blob handler to automatically expire stale
sessions after 10 minutes.

Not sure how I ran into this issue (unable to replicate), but I did have
an issue with max open sessions when stress testing on a machine. I hope
this change will improve reliability.

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: Ic302e8cb37cdbc9913f5b9ced004dabc27cac14d