Add unittest for unused delete and stat.
Since the documentation now specifies that this functions are meant to
always fail because they do not fit the current handler's design, I
thought the tests should also reflect this.

If the design ever changes and this are implemented (and the
documentation updated), then this tests will simply serve as a reminder
to add unittests for the new implementation (go TDD!).

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: Iefc20aa2be304a7bca222d8120870c2fbe3106e6