Remove lite runtime option from rrd proto.
This line was added in the original implementation since it is
recomended for devices with limited resources. However, the protobuf
package being used automatically links the full runtime, so we limit
ourself to the MessageLite interface for not a lot of gains.

I tried changing the PKG_CHECK_MODULE directive to use the protobuf-lite
package instead, but I discovered that the lite implementation does not
support well-known types[1]. Futhermore, this proto definition should be
shared with the host implementation, which does not have such a resource
constrain, but is not able to override it with build options.

If a workaround for the well-known types is implemented or discovered,
then we should update the build rules in the Makefile to enable a
runtime[2], instead of forcing everyone to a subset of functionality.

[1] https://github.com/protocolbuffers/protobuf/issues/1889
[2] https://github.com/protocolbuffers/protobuf/issues/487

Signed-off-by: Gabriel Matute <gmatute@google.com>
Google-Bug-Id: 137037681
Change-Id: I06ffed9d0951df5c1392cfc01831c072b67e5029